Pass all tests
For the last test to succeed the proc-macro
needed to be explicit about the correct types to be returned,
e.g. instead of simply returning a Result, in order to generate
"correct code" in each context, the proc macro needs to use the fully
qualified type std::result::Result so that it won't break in situations
where Result is redefined before the proc-macro is used.